requirements.txt
-i https://pypi.org/simple
attrs==22.2.0 ; python_version >= '3.6'
black==23.1.0
click==8.1.3 ; python_version >= '3.7'
exceptiongroup==1.1.0 ; python_version < '3.11'
flake8==6.0.0
iniconfig==2.0.0 ; python_version >= '3.7'
mccabe==0.7.0 ; python_version >= '3.6'
mypy-extensions==1.0.0 ; python_version >= '3.5'
packaging==23.0 ; python_version >= '3.7'
pathspec==0.11.0 ; python_version >= '3.7'
platformdirs==3.0.0 ; python_version >= '3.7'
pluggy==1.0.0 ; python_version >= '3.6'
pycodestyle==2.10.0 ; python_version >= '3.6'
pyflakes==3.0.1 ; python_version >= '3.6'
pytest==7.2.1
tomli==2.0.1 ; python_version < '3.11'
typing-extensions==4.5.0 ; python_version < '3.10'
numpy==1.24.2 ; python_version >= '3.7'
opencv-python-headless==4.7.0.68
pillow==9.4.0
